[UPD] loan models
