+++ loan_application/models/__init__.py
@@ -1,2 +1,4 @@
 from . import models
 from . import loan_application_document_type
+from . import loan_tag
+from . import loan_document
