+++ loan_application/models/models.py
@@ -2,47 +2,6 @@ from odoo import fields, models, api
 from odoo.exceptions import UserError, ValidationError
 from datetime import date
 
-class LoanDocument(models.Model):
-    _name = 'loan.document'
-    _description = 'Loan Document'
-    _order = "sequence, id"
-    
-
-    name = fields.Char(string='Name', required=True)
-    application_id = fields.Many2one('loan.application', string='Application')
-    attachment = fields.Binary(string='File')
-    type = fields.Char(string='Type')
-    state = fields.Selection([
-        ('new', 'New'),
-        ('approved', 'Approved'),
-        ('rejected', 'Rejected')
-    ], string='State', required=True, default='new')
-    sequence = fields.Integer(string='Sequence', default=10)
-
-    @api.onchange('attachment')
-    def _onchange_attachment(self):
-        for record in self:
-            if record.attachment:
-                record.state = 'new'
-
-    def action_approve(self):
-        for record in self:
-            record.write({'state': 'approved'})
-
-    def action_reject(self):
-        for record in self:
-            record.write({'state': 'rejected'})
-
-class LoanTag(models.Model):
-    _name = 'loan.tag'
-    _description = 'Loan Tag'
-    _order = "name"
-
-
-    name = fields.Char(string='Name', required=True)
-    color = fields.Integer(string='Color')
-
-    _sql_constraints = [('unique_tag_name', 'UNIQUE(name)', 'Tag name must be unique!')]
 
 class LoanApplication(models.Model):
     _name = "loan.application"
